all numbers are floats now

+++ src/app/ui/main.py
@@ -127,7 +127,7 @@ def generate_from_group(file):
                     # Fill the numerical fields
                     if(columns_num.__contains__(field_name)) :
                         try:
-                            data_num[field_name] = int(field_value.value)
+                            data_num[field_name] = float(field_value.value)
                         except ValueError:
                             data_num[field_name] = 0
